[Mod] split rpc module into client.py and server.py
